% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_WRAP.R
\name{add_phase}
\alias{add_phase}
\title{Function to overlay phase information on data}
\usage{
add_phase(x, times, labels = NULL)
}
\arguments{
\item{x}{a `sleep`, `activity`, or `stress` object produced by `read_wrap`}

\item{times}{a vector of UTC times that will serve as cutpoints an the `startTimeInSeconds`
variable. There should be one fewer `times` than the desired number of phase
to be produced.}

\item{labels}{a vector of the same length as `times` containing labels for the
phases. Optional. If `labels` is `NULL` the phase labels will be 1, 2, 3, etc.}
}
\description{
Function to overlay phase information on data
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
