% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuite.R
\name{adjuster}
\alias{adjuster}
\title{adjuster - a function to adjust p-values for a given column `wc` in the data
`x`.}
\usage{
adjuster(x, method = "BY")
}
\description{
note that with missing data, the parameter `n` in p.adjust will be
length(temp), not length(na.omit(temp)). If the latter is used, results
will be more conservative, i.e., adjusting for the number of converged
analyses rather than the total number of analyses
}
