% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{alf}
\alias{alf}
\title{alf - attach(as.list(formals(x))), a debugging tool}
\usage{
alf(x)
}
\description{
alf - attach(as.list(formals(x))), a debugging tool
}
\author{
Stephen Tueller \email{Stueller@rti.org}
}
\keyword{internal}
