% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{alignPhases}
\alias{alignPhases}
\title{align the time variable to be zero at the transition between 
the first and second phase}
\usage{
alignPhases(dat, id, phase, time, do.plot = FALSE)
}
\arguments{
\item{id}{See \code{\link{PersonAlytic}}.}

\item{phase}{See \code{\link{PersonAlytic}}.}

\item{time}{See \code{\link{PersonAlytic}}.}

\item{do.plot}{Logical. Should the resulting data be plotted?}

\item{data}{See \code{\link{PersonAlytic}}.}
}
\description{
align the time variable to be zero at the transition between 
the first and second phase
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
