% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_WRAP.R
\name{append_dailies}
\alias{append_dailies}
\title{Function to add daily data to WRAP data for a list of daily variables}
\usage{
append_dailies(data, wrap, dailies)
}
\arguments{
\item{data}{A data frame in wide format containing multiple daily variables.
This data must include an ID variable named `id` with integer values starting
at 1 to match the default ID variable naming used by WRAP.}

\item{wrap}{WRAP data produced by `read_wraps` (e.g., sleep, activity, or stress)}

\item{dailies}{A list of daily variable names 
`e.g., list(c(x_1, x_2, ..., x_t), c(y_1, y_2, ..., y_2))` where names are in 
the format `varName_#`, where `#` are integers that overlap with the range of 
values in the WRAP data variable `time_days`, e.g. `range(sleep$time_days)`.}
}
\description{
Function to add daily data to WRAP data for a list of daily variables
}
