% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palytic.r
\name{ARpq}
\alias{ARpq}
\title{ARpq - helper function for $GroupAR_order}
\usage{
ARpq(clone, p, q, subgroup)
}
\description{
ARpq - helper function for $GroupAR_order
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
