% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{autoDetect}
\alias{autoDetect}
\title{autoDetect}
\usage{
autoDetect(t0, userFormula, dims, detectTO, detectAR, maxOrder, whichIC,
  PQ, doForeach = TRUE)
}
\description{
autoDetect
}
\keyword{internal}
