% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterize.r
\name{byPhasebyGroup}
\alias{byPhasebyGroup}
\title{Function to create dummy codes for phase and group specific effects.}
\usage{
byPhasebyGroup(data, time, phase, group)
}
\arguments{
\item{data}{A \code{data.frame} with a time and phase variables
(see \code{\link{PersonAlytic}}), and a grouping variable (which would
normally be passed to \code{PersonAlytic} via the \code{ivs} parameter).}

\item{time}{Character. The name of the time variable.}

\item{phase}{Character. The name of the phase variable.}

\item{group}{Character. The name of the group variable.}
}
\description{
In ICT data with 2+ groups and 2+ phases, dummy coding can be used to
estimate group and phase specific intercepts and slopes.
}
\examples{

# create group and phase specific dummy indicators after creating an
# artificial "Group" variable in the Ovary data.
OvaryICT2 <- OvaryICT
OvaryICT2$Group <- as.numeric(OvaryICT2$Mare<7)
OvaryICT2 <- byPhasebyGroup(OvaryICT2, "Time", "Phase", "Group")

# view the new variables, group and phase specific intercepts and slopes
OvaryICT2$dummyNames

# view the right hand side formula of the fixed effects produced by
# `byPhasebyGroup`, noting the `-1` elimination of an overall intercept
OvaryICT2$fixed

# fit a model using the new variables with
# - no intercept
# - no overall time variable
userFormula <- list()
userFormula$fixed <- formula(paste("follicles ~ ", OvaryICT2$fixed)) # add dependent variable
userFormula$random <- ~ Time | Mare

t0 <- PersonAlytic(output      = 'byPhasebyGroup'     ,
                   data        = OvaryICT2$data       ,
                   ids         = "Mare"               ,
                   dvs         = "follicles"          ,
                   time        = "Time"               ,
                   ivs         = OvaryICT2$dummyNames ,
                   package     = "nlme"               ,
                   userFormula = userFormula          ,
                   correlation = "corARMA(p=1)"       ,
                   autoSelect  = list()               ,
                   method      = "ML"                 )
summary(t0)

# note that the following moves the `-1` to the end of the formula, test
# that this still works
userFormula$fixed <- update(formula(paste(" ~ ", OvaryICT2$fixed)), follicles ~ .)
t1 <- PersonAlytic(output      = 'byPhasebyGroup'     ,
                   data        = OvaryICT2$data       ,
                   ids         = "Mare"               ,
                   dvs         = "follicles"          ,
                   time        = "Time"               ,
                   ivs         = OvaryICT2$dummyNames ,
                   package     = "nlme"               ,
                   userFormula = userFormula          ,
                   correlation = "corARMA(p=1)"       ,
                   autoSelect  = list()               ,
                   method      = "ML"                 )
summary(t1)

}
