% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{checkID}
\alias{checkID}
\title{Function to check ID variable and if possible, coerce to numeric}
\usage{
checkID(x, ids)
}
\arguments{
\item{x}{The ID variable values as a vector.}

\item{ids}{The ID variable name. See \code{\link{PersonAlytic}}.}
}
\description{
Function to check ID variable and if possible, coerce to numeric
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
