% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean}
\alias{clean}
\title{clean, a function to check data and inputs.}
\usage{
clean(data, ids, dv, time, phase = NULL, ivs = NULL, fixed = NULL,
  random = NULL, formula = NULL, correlation = NULL, family = NULL,
  dvs = NULL, target_ivs = NULL, standardize = list(dv = FALSE, iv =
  FALSE, byids = FALSE), sortData = TRUE, alignPhase = "none",
  debugforeach = debugforeach)
}
\arguments{
\item{data}{See \code{\link{PersonAlytic}}.}

\item{ids}{See \code{\link{PersonAlytic}}.}

\item{dv}{See \code{\link{Palytic}}.}

\item{time}{See \code{\link{PersonAlytic}}.}

\item{phase}{See \code{\link{PersonAlytic}}.}

\item{ivs}{See \code{PersonAlytic}.}

\item{fixed}{See \code{\link{Palytic}}.}

\item{random}{See \code{\link{Palytic}}.}

\item{formula}{See \code{\link{Palytic}}.}

\item{correlation}{See \code{\link{PersonAlytic}}.}

\item{family}{See \code{\link{PersonAlytic}}.}

\item{dvs}{See \code{\link{PersonAlytic}}.}

\item{target_ivs}{See \code{PersonAlytic}.}

\item{standardize}{Logical. Should all variables be standardized? Only applies
to \code{dvs}, \code{ivs}, and \code{target_ivs}.}

\item{alignPhase}{See \code{\link{PersonAlytic}}.}
}
\description{
clean, a function to check data and inputs.
}
\examples{
\dontrun{
data <- clean(data=OvaryICT, ids='Mare', dv='follicles',
              time='Time', phase='Phase')
# illustrate dropping cases with < 2 observations
data <- clean(data=OvaryICT[29:nrow(OvaryICT),], ids='Mare', dv='follicles',
              time='Time', phase='Phase')
}

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
