% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_curelator.r
\name{clean_curelator}
\alias{clean_curelator}
\title{clean_curelator, a function to clean Curelator input dataa.}
\usage{
clean_curelator(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}} with ---specs here---.}
}
\description{
To be moved to package with Curelator specific tools (this may be
the only thing in it).
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
