% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cleanCall}
\alias{cleanCall}
\title{cleanCall - clean up the call in in palytic objects}
\usage{
cleanCall(modelResult, PalyticObj)
}
\description{
cleanCall - clean up the call in in palytic objects
}
\keyword{internal}
