% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTests.R
\name{corTests}
\alias{corTests}
\title{corTests -- this should be turned into a testthat test}
\usage{
corTests()
}
\arguments{
\item{x}{Any formula}
}
\description{
corTests -- this should be turned into a testthat test
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
