% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{data}
\name{CurEx}
\alias{CurEx}
\title{Curelator test data}
\description{
Curelator test data
}
\examples{
# no predictor model
CurExMod <- PersonAlytic(data=CurEx,
                 ids="id",
                 dv="severity.migraine",
                 time="time",
                 time_power=3,
                 standardize=TRUE,
                 package="nlme")
summary(CurExMod)
# angriness predictor model
CurExModAng <- PersonAlytic(data=CurEx,
                 ids="id",
                 dv="severity.migraine",
                 time="time",
                 ivs=list('angriness'),
                 time_power=3,
                 standardize=TRUE,
                 package="nlme")
summary(CurExModAng)
# angriness predictor model, add correlation
CurExModAngAR1 <- PersonAlytic(data=CurEx,
                 ids="id",
                 dv="severity.migraine",
                 time="time",
                 ivs=list('angriness'),
                 time_power=3,
                 correlation="corARMA(p=1)",
                 standardize=TRUE,
                 package="nlme")
summary(CurExModAngAR1)

# demonstrate how random effects modeling components affect bias and precision
# in eastimating the mean of a repeated measures variable with n=1
CurEx.mn <- mean(CurEx$severity.migraine)
CurEx.sd <- sd(CurEx$severity.migraine)/sqrt(nrow(CurEx))
CurEx.lm <- lm(severity.migraine~1, data=CurEx)
CurEx.arm <- ar(CurEx$severity.migraine)$x.mean
CurEx.ars <- ar(CurEx$severity.migraine)$var.pred
CurEx.ri <- nlme::lme(severity.migraine~1, data=CurEx, random=~1|id)
CurEx.rs <- nlme::lme(severity.migraine~1, data=CurEx, random=~time|id)
CurEx.rsar1 <- nlme::lme(severity.migraine~1, data=CurEx,
                         random=~time|id, correlation=nlme::corAR1())
CurEx.rst <- nlme::lme(severity.migraine~time, data=CurEx, random=~time|id)
CurEx.rstar1 <- nlme::lme(severity.migraine~time, data=CurEx, random=~time|id,
                          correlation=nlme::corAR1())
rownms <- c('Descriptive', 'Linear Model', 'Autoregressive Model',
            'Random Intercept', 'Random Intercept & Slope',
            'Random Intercept & Slope  & AR(1)',
            'Random Intercept & Slope & Fixed Effects for Time',
            'Random Intercept & Slope & AR(1) & Fixed Effects for Time')
outdf <- data.frame(rownms,
                    matrix(c(CurEx.mn, CurEx.sd,
                             summary(CurEx.lm)$coef[1:2],
                             CurEx.arm, CurEx.ars,
                             summary(CurEx.ri)$tTable[1:2],
                             summary(CurEx.rs)$tTable[1:2],
                             summary(CurEx.rsar1)$tTable[1:2],
                             summary(CurEx.rst)$tTable[1,1:2],
                             summary(CurEx.rstar1)$tTable[1,1:2]), ncol=2, byrow = TRUE)
)
names(outdf) <- c('Model', 'Mean', 'se')
outdf
}
\author{
Stephen Tueller \email{stueller@rti.org}

Example data from Curelator
}
