% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{decompFormula}
\alias{decompFormula}
\title{decompFormula - function to deconstruct a return formula}
\usage{
decompFormula(formula = NULL)
}
\arguments{
\item{See}{\code{\link{Palytic}}}
}
\description{
decompFormula - function to deconstruct a return formula
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
