% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palytic.r
\name{.active}
\alias{.active}
\title{.active - active bindings for Palytic objects}
\usage{
.active()
}
\description{
.active - active bindings for Palytic objects
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
