% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{.htp}
\alias{.htp}
\title{.htp}
\usage{
.htp(
  t1,
  id,
  iv,
  dv,
  dvs,
  ivs,
  dims,
  package,
  target_ivs,
  PQ,
  family,
  fpc,
  popsize2,
  debugforeach
)
}
\description{
.htp
}
\keyword{internal}
