% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dstats}
\alias{dstats}
\title{dstats - print mean, median, sd, skewness, and kurtosis - 3}
\usage{
dstats(dv, phase = NULL, more = FALSE, print = FALSE)
}
\description{
dstats - print mean, median, sd, skewness, and kurtosis - 3
}
