% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eds.r
\name{eds}
\alias{eds}
\title{simplify exists(deparse(substitute(x)))}
\usage{
eds(x)
}
\arguments{
\item{x.}{The name of any type of object}
}
\value{
Logical. Does the object exist?
}
\description{
simplify exists(deparse(substitute(x)))
}
