% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{fitWithTargetIV}
\alias{fitWithTargetIV}
\title{fit models with the target iv & calculate LRT}
\usage{
fitWithTargetIV(t0, package, useObs, dims, mod1 = NULL, PQ = c(3, 3))
}
\description{
fit models with the target iv & calculate LRT
}
\keyword{internal}
