% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{fitWithTargetIVarma}
\alias{fitWithTargetIVarma}
\alias{fitWithTargetIVarma}
\title{fitWithTargetIVarma}
\usage{
fitWithTargetIVarma(t1, useObs, dims, PQ)

fitWithTargetIVarma(t1, useObs, dims, PQ)
}
\description{
fitWithTargetIVarma

fitWithTargetIVarma
}
\keyword{internal}
