% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{fixPhaseNA}
\alias{fixPhaseNA}
\title{fixPhaseNA - function to carry last value forward if phase is missing
(this is common in example data sets Ty has worked with)}
\usage{
fixPhaseNA(phase)
}
\arguments{
\item{phase}{The \code{phase} variable}
}
\description{
fixPhaseNA - function to carry last value forward if phase is missing
(this is common in example data sets Ty has worked with)
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
