% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{forms}
\alias{forms}
\title{forms - function to construct and return formulae (fixed, random, formula)
whenever a change to any variable that can go into a formula is made}
\usage{
forms(
  data,
  PalyticObj = NULL,
  ids = NULL,
  dv = NULL,
  time = NULL,
  phase = NULL,
  ivs = NULL,
  interactions = NULL,
  time_power = NULL,
  correlation = NULL,
  family = NULL,
  fixed = NULL,
  random = NULL,
  formula = NULL,
  method = "REML",
  dropTime = "no",
  corFromPalyticObj = TRUE
)
}
\arguments{
\item{See}{\code{\link{Palytic}}}
}
\description{
Note that only the following combinations work, but these are enforced in
\code{\link{Palytic}} rather than enforced internally
1. Objects except for PalyticObj, fixed, random, and formula (e.g., Palytic$new)
2. PalyticObj and exactly 1 other parameter
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
