% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovFPC.R
\name{FPC}
\alias{FPC}
\title{FPC function to get the finite population corrected standard errors and
update the tTable of an lme object}
\usage{
FPC(object, popsize2)
}
\arguments{
\item{object}{An \code{\link{lme}} object from \code{\link{nlme}}, or a merMod
object from \code{\link{lmer}}.}

\item{popsize2}{The finite population size at level two, i.e., persons when
there are repeated measures at level 1.}
}
\description{
FPC function to get the finite population corrected standard errors and
update the tTable of an lme object
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
