% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{frmToChar}
\alias{frmToChar}
\title{frmToChar - convert formula to text}
\usage{
frmToChar(x)
}
\arguments{
\item{x}{Any formula}
}
\description{
frmToChar - convert formula to text
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
