% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{gerARIMAorder}
\alias{gerARIMAorder}
\alias{gerARIMAorder}
\title{gerARIMAorder}
\usage{
gerARIMAorder(x)

gerARIMAorder(x)
}
\description{
gerARIMAorder

gerARIMAorder
}
\keyword{internal}
