% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{getARMAtbl}
\alias{getARMAtbl}
\alias{getARMAtbl}
\title{getARMAtbl: get ARMA tTable}
\usage{
getARMAtbl(x)

getARMAtbl(x)
}
\description{
getARMAtbl: get ARMA tTable

getARMAtbl: get ARMA tTable
}
\keyword{internal}
