% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitations.r
\name{getCitations}
\alias{getCitations}
\title{Citations for PersonAlytics}
\usage{
getCitations(load = FALSE)
}
\arguments{
\item{load}{Logical. Should packages be loaded? If \code{FALSE} (the default) citations are returned as a list. Called internally by [link to function].}
}
\description{
Citations for PersonAlytics
}
\examples{
getCitations()

}
