% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovFPC.R
\name{getLambdat}
\alias{getLambdat}
\title{getLambdat - get Lambdat from an lme object analogous to
getME(object.merMod, "Lambdat"). See \code{\link{getMe}}.}
\usage{
getLambdat(object)
}
\arguments{
\item{object}{An \code{lme} model}
}
\description{
getLambdat - get Lambdat from an lme object analogous to
getME(object.merMod, "Lambdat"). See \code{\link{getMe}}.
}
\examples{

# see examples in \\code{vcovFPC}
}
