% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalytic.R
\name{getPalytic}
\alias{getPalytic}
\title{getPalytic, a function to create a Palytic object from PersonAlytics output}
\usage{
getPalytic(
  output,
  data,
  rowNum,
  id = NULL,
  dv = NULL,
  target_iv = NULL,
  ids = NULL
)
}
\arguments{
\item{output}{Character. The path to a csv file produced by the \code{PersonAlytic} function or
the data.frame returned by the \code{PersonAlytic} function.}

\item{data}{Data Frame. The same data provided to your \code{PersonAlytic} run.}

\item{rowNum}{Numeric. The row number you want to create a \code{Palytic} object for. Set to
\code{NULL} if \code{id}, \code{dv}, \code{target_iv}, and \code{ids}
will be provided, noting that all four of these parameters must be specified except
\code{target_iv}, but only if \code{target_iv} is not a column in \code{output}.}

\item{id}{Character. The value in the 'id' column of \code{output}.}

\item{dv}{Character. The value in the 'dv' column of \code{output}.}

\item{target_iv}{Character. The value in the 'target_iv' column of \code{output}. Leave as
\code{NULL} if this column is absent in \code{output}.}
}
\description{
getPalytic, a function to create a Palytic object from PersonAlytics output
}
\examples{
\dontrun{

t1 <- PersonAlytic(output          = 'Test1'     ,
                   data            = OvaryICT    ,
                   ids             = "Mare"      ,
                   dvs             = "follicles" ,
                   phase           = "Phase"     ,
                   time            = "Time"      ,
                   package         = "arma"      ,
                   individual_mods = TRUE        )
 Mare1 <- getPalytic('Test1_PersonAlytic.csv', data=OvaryICT, rowNum=1)

}

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
