% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{getParameters}
\alias{getParameters}
\title{getParameters: extract parameter table, the `Model` variable is too big}
\usage{
getParameters(Model, package, target_iv)
}
\description{
getParameters: extract parameter table, the `Model` variable is too big
}
\keyword{internal}
