% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovFPC.R
\name{getZt}
\alias{getZt}
\title{getZt - get Zt from an lme object analogous to
getME(object.merMod, "Zt"). See \code{\link{getME}}.}
\usage{
getZt(object)
}
\arguments{
\item{object}{An \code{lme} model}
}
\description{
getZt - get Zt from an lme object analogous to
getME(object.merMod, "Zt"). See \code{\link{getME}}.
}
\examples{

# see examples in \code{vcovFPC}
}
