% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.foreach.r
\name{htp.foreach}
\alias{htp.foreach}
\title{high throughput}
\usage{
htp.foreach(data, dims, dvs, phase, ids, uids, time, ivs, target_ivs,
  interactions = NULL, time_power = 1, correlation = NULL,
  family = gamlss.dist::NO(), standardize = TRUE, package = "gamlss",
  detectAR = TRUE, PQ = c(3, 3), whichIC = "BIC", detectTO = TRUE,
  maxOrder = 3, sigma.formula = ~1, debugforeach = FALSE)
}
\description{
high throughput
}
\keyword{internal}
