% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{htp}
\alias{htp}
\title{high throughput}
\usage{
htp(
  data,
  dims,
  dvs,
  phase,
  ids,
  uids,
  time,
  ivs,
  target_ivs,
  interactions = NULL,
  time_power = 1,
  alignPhase = "none",
  correlation = NULL,
  family = gamlss.dist::NO(),
  standardize = list(dvs = FALSE, ivs = FALSE, byids = FALSE),
  fpc = FALSE,
  popsize2 = 0,
  package = "gamlss",
  autoSelect = list(AR = list(P = 3, Q = 3), TO = list(polyMax = 3), DIST = list()),
  whichIC = "BIC",
  sigma.formula = ~1,
  debugforeach = FALSE,
  userFormula = list(fixed = NULL, random = NULL, formula = NULL),
  cores = parallel::detectCores() - 1
)
}
\description{
high throughput
}
\keyword{internal}
