% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{htpErrors}
\alias{htpErrors}
\alias{htpErrors}
\title{htpErrors: accumulate inputs and errors}
\usage{
htpErrors(t1, id, dv, dims, package, useObs, target_iv)

htpErrors(t1, id, dv, dims, package, useObs, target_iv)
}
\description{
htpErrors: accumulate inputs and errors

htpErrors: accumulate inputs and errors
}
\keyword{internal}
