% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ICC}
\alias{ICC}
\title{ICC function from \href{http://davidakenny.net/papers/k&h/R_output.txt}{David Kenny}}
\usage{
ICC(out)
}
\arguments{
\item{out}{An \code{lme} object}
}
\description{
ICC function from \href{http://davidakenny.net/papers/k&h/R_output.txt}{David Kenny}
}
\examples{
egaov <- aov(follicles ~ factor(Mare), data = OvaryICT)
multilevel::ICC1(egaov)
multilevel::ICC2(egaov)
eglme <- nlme::lme(follicles ~ 1, data = OvaryICT, random = ~ 1 | Mare,
                   method = 'ML')
# VarCorr(eglme)
ICC(eglme)
library(gamlss)
eggamlss1 <- gamlss(follicles ~ 1 + re(random = ~ 1 | Mare, method = 'ML'),
                           data = OvaryICT,
                           sigma.formula = ~ 1)
# VarCorr(getSmo(eggamlss1))
ICC(eggamlss1)
eggamlss2 <- gamlss(follicles ~ 1 + re(random = ~ 1 | Mare, method = 'ML'),
                           data = OvaryICT,
                           sigma.formula = ~ 0)
# VarCorr(getSmo(eggamlss2))
ICC(eggamlss2)
}
\author{
\href{http://davidakenny.net}{David Kenny}
}
