% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTplot.r
\name{ICTplot}
\alias{ICTplot}
\title{ICTplot - function to plot ICT data.}
\usage{
ICTplot(self, data, legendName = NULL, type = c("density", "histogram",
  "freqpoly"), ylim = NULL)
}
\arguments{
\item{out}{An \code{lme} object}
}
\description{
ICTplot - function to plot ICT data.
}
\examples{

t1 <- Palytic$new(data = OvaryICT, ids='Mare', dv='follicles',
        time='Time', phase='Phase')
t1$plot()
t1$plot(groupvar = 'Target1')
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
