% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{isNullOrForm}
\alias{isNullOrForm}
\alias{isNullOrForm}
\title{isNullOrForm: is an object non-NULL and a formula?}
\usage{
isNullOrForm(x)

isNullOrForm(x)
}
\description{
isNullOrForm: is an object non-NULL and a formula?

isNullOrForm: is an object non-NULL and a formula?
}
