% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmee.r
\name{lmee}
\alias{lmee}
\title{Function to fit lme() models with error handling and automatically refitting using optim if nlminb fails to converge.}
\usage{
lmee(fixed, data, random, correlation = NULL, method = "REML", ...)
}
\arguments{
\item{fixed}{Formula. See \code{\link{lme}}.}

\item{data}{data.frame. See \code{\link{lme}}.}

\item{random}{Formula. See \code{\link{lme}}.}

\item{correlation}{Correlation structure class. \code{\link{lme}} and \code{\link{corClasses}}.}

\item{method}{Logical See ?\code{\link{lme}}.}
}
\value{
Either a \code{lme} object or a character message that the model did not converge.
}
\description{
Function to fit lme() models with error handling and automatically refitting using optim if nlminb fails to converge.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
