% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makeForms}
\alias{makeForms}
\title{makeForms - function to construct and return formulae (fixed, random, formula)}
\usage{
makeForms(ids = "Mare", dv = "follicles", time = "Time",
  phase = "Phase", ivs = "iv1", interactions = list(c("iv1", "Phase")),
  rint = "", correlation = "NULL", family = NO(), dropTime = FALSE,
  method = "REML")
}
\arguments{
\item{See}{\code{\link{Palytic}}}
}
\description{
makeForms - function to construct and return formulae (fixed, random, formula)
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
