% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{messenger}
\alias{messenger}
\title{messenger}
\usage{
messenger(
  dvLoop,
  dvs = NULL,
  dv = NULL,
  dims = NULL,
  ids = NULL,
  target_ivs = NULL
)
}
\description{
messenger
}
\keyword{internal}
