% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{monotone}
\alias{monotone}
\title{monotone}
\usage{
monotone(ids, time, data)
}
\arguments{
\item{x}{Any object}
}
\description{
monotone
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
