% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{nullString}
\alias{nullString}
\alias{nullString}
\title{nullString - turn NULL to "NULL"}
\usage{
nullString(x)

nullString(x)
}
\description{
nullString - turn NULL to "NULL"

nullString - turn NULL to "NULL"
}
\keyword{internal}
