% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytic.r
\name{paHTP}
\alias{paHTP}
\title{paHTP}
\usage{
paHTP(file, data, ids, dvs, time, phase, ivs, target_ivs, interactions,
  time_power, correlation, family, subgroup, standardize, package, ind.mods,
  PalyticObj, detectAR, PQ, IC, detectTO, charSub, sigma.formula, p.method,
  alpha, debugforeach = FALSE)
}
\description{
paHTP
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
