% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlyticHTP.r
\name{PersonAlyticHTP}
\alias{PersonAlyticHTP}
\title{High through-put options for Palytic objects.}
\usage{
PersonAlyticHTP(file = NULL, data = NULL, ids, dvs, time, phase = NULL,
  ivs = NULL, ivsl = NULL, interactions = NULL, time_power = 1,
  correlation = NULL, family = gamlss.dist::NO(), subgroup = NULL,
  standardize = TRUE, package = "gamlss", ind.mods = TRUE,
  grp.mod = FALSE, PalyticObj = NULL, detectAR = TRUE, detectTO = TRUE,
  maxOrder = 3, charSub = NULL, sigma.formula = ~1,
  debugforeach = FALSE, p.method = "BY", alpha = 0.05)
}
\arguments{
\item{file}{The file name (or full path with `/` instead of `\`) where output should
be saved. If left \code{NULL}, the date and time will prefix `PersonAlyticHTP_Output.csv`.}

\item{data}{See \code{\link{PersonAlytic}}.}

\item{ids}{See \code{\link{PersonAlytic}}.}

\item{dvs}{A list of one or more character dependent variable names in \code{data}.
The linear mixed effects model \code{dvs[d] ~ phase + time + phase*time + ivsl[c] + ivs}
with random effects \code{~ time | ids[i]} will be fit to the data using
\code{\link{gamlss}}. The iterators \code{[d]}, \code{[d]}, and \code{[d]}
indicates the model will be fit for each combination of dependent variable in \code{dvs},
independent variable in \code{ivsl}, and each unique ID in \code{ids}
(overridden using \code{ind.mods=FALSE}) controlling for indepented variables in
\code{ivs}. For more options submit a \code{PalyticObj}.}

\item{time}{See \code{\link{PersonAlytic}}.}

\item{phase}{See \code{\link{PersonAlytic}}.}

\item{ivs}{See \code{\link{PersonAlytic}}, noting that the variables in \code{ivs}
cannot also be in \code{ivsl}.}

\item{interactions}{See \code{\link{PersonAlytic}}.}

\item{time_power}{See \code{\link{PersonAlytic}}. Ignored if \code{detectAR=TRUE}}

\item{correlation}{See \code{\link{PersonAlytic}}. Ignored if \code{detectTO=TRUE}}

\item{family}{See \code{\link{PersonAlytic}}.}

\item{subgroup}{See \code{\link{PersonAlytic}}.}

\item{standardize}{See \code{\link{PersonAlytic}}. The default is \code{TRUE}
which makes parameter estimate magnitudes comparable across individuals, outcomes in
\code{dvs}, and covariates in \code{ivls}.}

\item{package}{See \code{\link{PersonAlytic}}.}

\item{ind.mods}{Logical, defaults to \code{TRUE}. Should individual models be
fit for each ID?}

\item{grp.mod}{Logical, defaults to \code{FALSE}. Should a group level model be fit
across all IDs? If both \code{ind.mods} and \code{grp.mod} are \code{FALSE},
\code{grp.mod} will be changed to \code{TRUE}.}

\item{PalyticObj}{See \code{\link{Palytic}}. If \code{PalyticObj} is submitted
then only \code{dvs}, \code{ivsl}, \code{ind.mods}, and \code{grp.mod} will be
used. This allows users access to additional options including generalized linear
mixed effects models via the \code{family} option, user specified \code{correlation}
structures (in non-\code{NULL} this will override the automated correlation structure
search), and user specified models via \code{formula}.}

\item{detectAR}{Logical, defaults to \code{TRUE}. Should the autoregressive structure
be automatically detected? If the \code{time} variable is equally spaced, this is
done using the function \code{\link{forecast}}. If the \code{time} variable is not
equally spaced, this is done using likelihood ratio tests with mixed effects models
using the specified \code{package} under maximum likelihood. This is done separately
for each case in \code{ids} if \code{ind.mods=TRUE}.}

\item{detectTO}{Logical, defaults to \code{TRUE}. Should the \code{time_power} value
be automatically detected? Values from 1 to \code{maxOrder} will be tested using
likelihood ratio tests with mixed effects models using the specified \code{package}
under maximum likelihood. This is done separately for each case in \code{ids}
if \code{ind.mods=TRUE}.}

\item{maxOrder}{Numeric, defaults to 3. What is the highest order of \code{time}
that sholud be tested for both fixed and randmo effects, e.g., \code{time +
I(time^2)+...+I(time^maxOrder)}.}

\item{charSub}{list of paired character strings for character substitution.
If the names of the target predictors
in \code{ivsl} had to be edited to make valid variable names, this parameter allows
users put the illegal characters back in. For example, if the original variable name
was "17.00_832.2375m/z", a letter would need to prefix the variable name and the
"/" would need to be replaced with another character, e.g., "X17.00_832.2375m.z".
To get the row names of the output back to original varibale name, use
\code{charSub=list(c("X", ""), c("m.z", "m/z"))}. Note that inputs to charSub
must be in double quotes and are case sensitive. All duplicates will be substituted.
For example, if the variable name was "X1X23.x" and \code{charSub=list(c("X", ""))},
the resulting row label for this variable would be "123.x".}

\item{sigma.formula}{A formula for the variance under \code{\link{gamlss}}. Static.
It will not change dynamically over iterations nor will it be updated by \code{time_power}
or \code{detectTO}. If model fitting using this option fails, another attempt will be
made after reseting it to its defaul, i.e., \code{~1}.}

\item{ivls}{Independent variables that are iterated over one at a time. Effects for
these variables are labeled as 'target predictor' in the output.}
}
\description{
A user interface for creating a Palytic object
(which inherits from \code{\link{Palytic}} and invoking
high through-put options including automated autocorrelation detection, automated
detection of a polynomial for time, and high throughput analyses of all individuals
in the data set, multiple dependent variables, and target independent variables.
Type I error corrections are implemented via [pending].
}
\examples{
# group model
t1 <- PersonAlyticHTP(data=OvaryICT,
                 ids="Mare",
                 dvs="follicles",
                 phase="Phase",
                 time="TimeSin",
                 package='nlme',
                 ind.mods=FALSE,
                 grp.mod=TRUE)
# individual models (using defaults)
t1 <- PersonAlyticHTP(data=OvaryICT,
                 ids="Mare",
                 dvs="follicles",
                 phase="Phase",
                 time="TimeSin",
                 package='nlme')
hist(t1$Phase.Std.Error, main='Distribution of Phase Effect', xlab='Phase Effect')
\dontrun{
# if you wish to delete the automatically created csv file, run
#NOT IMPLEMENTED YET
}
}
