% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{package}
\name{PersonAlytics}
\alias{PersonAlytics}
\title{PersonAlytics: Analytics for single-case, small N, and Idiographic Clinical Trials}
\description{
The PersonAlytics package provides the simplified user interface
for implementing linear mixed effects models for idiographic clinical trials
(ICT) data, single case studies, and small N studies with intensive longitudinal
designs. Contact us via https://personalytics.rti.org/ for
licensing options.
}
\details{
The basic mixed effects model is \eqn{dv=time+phase+phase*time}
with random intercepts and random slopes for time. The phase variable is optional.
Additional independent variables (or covariates) can be included.
The \code{PersonAlytics} package provides the simplified user interface
for implementing this model using \code{\link{gamlss}} or \code{\link{lme}}. The
primary function of \code{PersonAlytics} is \code{\link{PersonAlytic}}.

Key features of the \code{PersonAlytics} package include:

\strong{Automated detection of the residual covariance structure.}
\code{PersonAlytics} automates model comparisons for determining autocorrelation
structure for all patients or for each patient.

\strong{Automated detection of the function form for the time variable.}
\code{PersonAlytics} automates model comparisons for determining the functional
form of the relationship between time and the outcome
(i.e., linear vs. quadratic vs. cubic growth models) for all patients or for each patient.

\strong{Estimation.} The automated covariance structure and function form for time is
done using maximum likelihood (ML) estimators. Final results are estimated using
restricted maximum likelihood (REML).

\strong{High Throughput.} When users have a list of outcomes (dependent variables),
a list of target covariates, and/or
or desire the analyses to be repeated for each individual in the data set, high
throughput options automate the model fitting process.

\strong{False Discovery Rate Adjustment.} When high throughput options are requested,
Type I error correction and false discovery rate adjustments are implemented
post-implementation across target covariates (and individuals if requested) within
each outcome.

\strong{Linear and Generalized Linear Mixed Effects Models.} Linear mixed effects models
can be fit in either the \code{\link{nlme}} framework or the \code{\link{gamlss}}
approach. The two approaches give nearly identical fixed effects estimates but differ
in their computation of standard errors and random effects. Generalized linear
mixed effects models can be fit using the \code{\link{gamlss}} option
(see \code{\link{gamlss.family}}). The \code{\link{gamlss}}  approach also allows models
for dealing with heteroscedasticity implemented by including mixed effects models for
the variance.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
