% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotICT.r
\name{plotICT}
\alias{plotICT}
\title{plotICT - function to plot ICT data.}
\usage{
plotICT(self, data, legendName = NULL, type = c("density", "histogram",
  "freqpoly"), ylim = NULL)
}
\arguments{
\item{out}{An \code{lme} object}
}
\description{
plotICT - function to plot ICT data.
}
\examples{

t1 <- Palytic$new(data = OvaryICT, ids='Mare', dv='follicles',
        time='Time', phase='Phase')
t1$plot()
t1$plot(groupvar = 'Target1')
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
