% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotICT.r
\name{plotICTraw}
\alias{plotICTraw}
\title{plotICTraw - function to plot ICT data.}
\usage{
plotICTraw(
  x,
  npg,
  id = "id",
  group = "group",
  Time = "Time",
  y = "y",
  seed = 2
)
}
\arguments{
\item{x}{A data set}

\item{npg}{The number of observations per group to sample for plotting}

\item{id}{The name of the id variable}

\item{group}{The name of the grouping variable, if there is no grouping
variable, set to NULL}

\item{Time}{The name of the time variable}

\item{y}{The name of the dependent variable}

\item{seed}{The random seed for selecting \code{npg} individuals per group
for sampling}
}
\description{
plotICTraw - function to plot ICT data.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
