% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{pstand}
\alias{pstand}
\title{Function to standardize data with options from the \code{standardize} parameter.}
\usage{
pstand(data, standardize, dv, ivs, family, ids)
}
\arguments{
\item{data}{See \code{\link{PersonAlytic}}.}

\item{standardize}{See \code{\link{PersonAlytic}}.}

\item{dv}{See \code{\link{PersonAlytic}}.}

\item{ivs}{See \code{\link{PersonAlytic}}.}

\item{family}{See \code{\link{PersonAlytic}}.}

\item{ids}{See \code{\link{PersonAlytic}}.}
}
\description{
Function to standardize data with options from the \code{standardize} parameter.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
