% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{rcm}
\alias{rcm}
\title{rcm: function to force values to data.frame vectors for later stacking}
\usage{
rcm(x, target_ivs, data)
}
\arguments{
\item{x}{is a matrix-like ANOVA-like table from lme, gamlss, or coeftest in
arima}

\item{target_ivs}{See \code{targe_ivs} in \code{\link{PersonAlytic}}}

\item{data}{A data.frame, for now only pass t1$datac in \code{\link{htp}}}
}
\description{
rcm: function to force values to data.frame vectors for later stacking
}
\keyword{internal}
