% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{rcn}
\alias{rcn}
\alias{rcn}
\title{rcn: function to prep row names}
\usage{
rcn(x, target_ivs, data)

rcn(x, target_ivs, data)
}
\arguments{
\item{x}{See `x` for \code{\link{rcm}}}

\item{target_ivs}{See \code{targe_ivs} in \code{\link{PersonAlytic}}}

\item{data}{A data.frame, for now only pass t1$datac in \code{\link{htp}}}
}
\description{
rcn: function to prep row names

rcn: function to prep row names
}
\keyword{internal}
