% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_WRAP.R
\name{read_wrap}
\alias{read_wrap}
\title{Read and extract json WRAP data}
\usage{
read_wrap(file, study_start = NULL)
}
\arguments{
\item{file}{a character file name for json WRAP data}

\item{study_start}{a start time for zeroing activity the time variables.}
}
\description{
Read and extract json WRAP data
}
