% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_WRAP.R
\name{read_wraps}
\alias{read_wraps}
\title{Read and extract json WRAP data from multiple participants}
\usage{
read_wraps(files, study_start = NULL, times = NULL, labels = NULL)
}
\arguments{
\item{times}{see \code{\link(add_phase)}.}

\item{labels}{see \code{\link(add_phase)}.}

\item{file}{a vector json file names from WRAP}
}
\description{
Read and extract json WRAP data from multiple participants
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
