% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{remove_terms}
\alias{remove_terms}
\title{remove_terms from https://stackoverflow.com/questions/23381616/r-update-function-how-to-drop-all-the-variables-that-are-related-to-a-pre-speci}
\usage{
remove_terms(form, term)
}
\arguments{
\item{form}{A formula}

\item{term}{A term to be removed from the formula}
}
\description{
remove_terms from https://stackoverflow.com/questions/23381616/r-update-function-how-to-drop-all-the-variables-that-are-related-to-a-pre-speci
}
\author{
Sven Hohenstein
}
\keyword{internal}
