% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp-RTI-102921.r, R/htp.r
\name{rmSpecChar}
\alias{rmSpecChar}
\alias{rmSpecChar}
\title{rmSpecChar - strip out special characters from strings}
\usage{
rmSpecChar(x)

rmSpecChar(x)
}
\description{
rmSpecChar - strip out special characters from strings

rmSpecChar - strip out special characters from strings
}
\keyword{internal}
