% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_WRAP.R
\name{study_starts}
\alias{study_starts}
\title{Function to get the earliest time from a set of json WRAP files}
\usage{
study_starts(files)
}
\arguments{
\item{files}{a list of json file names. Use full path if files are not in
the working directory.}
}
\description{
Function to get the earliest time from a set of json WRAP files
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
