% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{subCheck}
\alias{subCheck}
\title{subCheck - test whether subgroup is the right length}
\usage{
subCheck(subgroup, data)
}
\arguments{
\item{subgroup}{logical vector on NULL}

\item{data}{A data set}
}
\description{
subCheck - test whether subgroup is the right length
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
