% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{subdat}
\alias{subdat}
\title{subdat - function to take a subset of data with column select}
\usage{
subdat(data, subgroup, formula)
}
\arguments{
\item{data}{data.frame.}

\item{subgroup}{Logical vector.}

\item{formula}{a formula.}
}
\description{
subdat - function to take a subset of data with column select
}
\author{
Stephen Tueller \email{Stueller@rti.org}
}
\keyword{internal}
