% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{summarySE}
\alias{summarySE}
\title{summarySE - function to take a subset of data with column select, used by
Palytic$plot}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, phase = NULL,
  na.rm = FALSE, conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{measurevar}{the name of a column that contains the variable to be summarized}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95\%)}
}
\description{
summarySE - function to take a subset of data with column select, used by
Palytic$plot
}
\author{
Updated by Stephen Tueller \email{Stueller@rti.org} from
\url{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)}
}
\keyword{internal}
