% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPsum.R
\name{sumPsum}
\alias{sumPsum}
\title{sumPsum, a function to summarize multiple p-value summaries}
\usage{
sumPsum(wd)
}
\arguments{
\item{wd}{A path to a directory containing multiple subdirectories, each with
one *.txt file produced by the \code{\link{psuite}} function.}
}
\value{
A *.csv file with the names of subfolders of the ‘run’ folders
in the first column, the proportion of unadjusted p values < .05 in the
second column, and the proportion of adjusted p values < .05 in the third column.

#' @author Stephen Tueller \email{stueller@rti.org}
}
\description{
sumPsum, a function to summarize multiple p-value summaries
}
