% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterize.r
\name{termsToFormula}
\alias{termsToFormula}
\title{Function to create a formula from a list of group and phase specific
intercepts and slopes}
\usage{
termsToFormula(terms, random = ~Time | id)
}
\arguments{
\item{terms}{Character vector or list.}

\item{random}{A random effects formula. The default is \code{~ Time | id} as
used by PersonAlyticsPower}
}
\description{
used by the PersonAlyticsPower GUI
}
\examples{

terms <- c("group1_phase1_int",
           "group1_phase1_slope",
           "group1_phase2_int",
           "group1_phase2_slope",
           "group1_phase3_int",
           "group1_phase3_slope",
           "group2_phase1_int",
           "group2_phase1_slope",
           "group2_phase2_int",
           "group2_phase2_slope",
           "group2_phase3_int",
           "group2_phase3_slope")
termsToFormula(terms)

}
