% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{to01}
\alias{to01}
\title{to01 - function to convent any variable to the [0,1] range}
\usage{
to01(x, na.rm = TRUE, squeeze = FALSE, prior = 0.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, should missing data be excluded when calculating min/max}
}
\description{
to01 - function to convent any variable to the [0,1] range
}
\references{
Smithson, M., & Verkuilen, J. (2006). A better lemon squeezer? Maximum-likelihood regression with beta-distributed dependent variables. Psychological Methods, 11(1), 54.
}
\author{
Stephen Tueller \email{Stueller@rti.org}
}
