% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuite.R
\name{trajplot}
\alias{trajplot}
\title{trajplot - plot trajectories}
\usage{
trajplot(data, ids, dv, time, phase, ivs, target_iv, target_nm)
}
\arguments{
\item{data}{Inputs from psuite()}

\item{ids}{See \code{\link{PersonAlytics}}}

\item{dv}{See \code{\link{PersonAlytics}}}

\item{time}{See \code{\link{PersonAlytics}}}

\item{phase}{See \code{\link{PersonAlytics}}}

\item{ivs}{See \code{\link{PersonAlytics}}}

\item{target_iv}{See \code{\link{PersonAlytics}}}

\item{target_nm}{The name of the target_iv}
}
\description{
trajectoy plots
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
