% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovFPC.R
\name{vcovFPC}
\alias{vcovFPC}
\alias{vcovFPC.meMod}
\alias{vcovFPC.merMod}
\alias{vcovFPC.lme}
\title{\code{vcovFPC} Obtain finite-population-adjusted standard errors for
fixed effects estimates for a fitted multilevel model}
\usage{
vcovFPC(object, ...)

\method{vcovFPC}{merMod}(object, popsize2 = NULL, popsize1 = NULL,
  KR = FALSE)

\method{vcovFPC}{lme}(object, popsize2 = NULL, popsize1 = NULL,
  KR = FALSE)
}
\arguments{
\item{object}{an R object of class merMod as resulting from lmer() or
an object of class lme resulting from lme().}

\item{popsize2}{population size at Level-2; if NULL, an infinite Level-2
population is assumed.}

\item{popsize1}{population size at Level-1; if NULL, an infinite Level-1
population is assumed.}

\item{KR}{Whether Kenward-Roger approximation of standard errors should be used,
which is recommended for small number of clusters and average cluster size.
Default to FALSE. Not available for lme objects.}
}
\value{
The variance-covariance matrix of the fixed effect estimates, as
returned by vcov()
}
\description{
\code{vcovFPC} Obtain finite-population-adjusted standard errors for
fixed effects estimates for a fitted multilevel model
}
\examples{

library(nlme)
library(lme4)

mod.lme <- lme(follicles ~ Time*Phase, data = OvaryICT, random = ~ Time | Mare,
               method = 'ML')
mod.merMod <- lmer(follicles ~ Time*Phase + (Time | Mare), data = OvaryICT,
                   REML = FALSE)

tTable.lme <- data.frame( summary(mod.lme)$tTable[,1:2] )
tTable.merMod <- data.frame( coef(summary(mod.merMod))[,1:2] )

fpc.lme <- vcovFPC(mod.merMod, popsize2 = 100)
fpc.merMod <- vcovFPC(mod.lme, popsize2 = 100)

tTable.lme$FPC.Std.Error <- sqrt(diag(fpc.lme))
tTable.merMod$FPC.Std.Error <- sqrt(diag(fpc.merMod))

tTable.lme
tTable.merMod


}
\references{
https://psycnet.apa.org/doiLanding?doi=10.1037%2Fmet0000137
}
\author{
Mark H. C. Lai \email{mark.lai@uc.edu}, Oi-man Kwok, Yu-Yu Hsiao,
and Quina Cao. Updated by Stephen Tueller \email{Stueller@rti.org} for
compatibility with \code{nlme}
for inclusion in \code{\link{PersonAlytics}}.
}
