% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean}
\alias{clean}
\title{clean, a function to check data and inputs.}
\usage{
clean(data, ids, dv, time, phase = NULL, ivs = NULL, fixed = NULL,
  random = NULL, formula = NULL, correlation = NULL, dvs = NULL,
  target_ivs = NULL, standardize = FALSE, sortData = TRUE,
  alignPhase = TRUE)
}
\arguments{
\item{data}{See \code{\link{PersonAlytic}}.}

\item{ids}{See \code{\link{PersonAlytic}}.}

\item{dv}{See \code{\link{PersonAlytic}}.}

\item{time}{See \code{\link{PersonAlytic}}.}

\item{phase}{See \code{\link{PersonAlytic}}.}

\item{ivs}{Used by \code{PersonAlyticPro}.}

\item{fixed}{See \code{\link{Palytic}}.}

\item{random}{See \code{\link{Palytic}}.}

\item{formula}{See \code{\link{Palytic}}.}

\item{dvs}{Used by \code{PersonAlyticPro}.}

\item{target_ivs}{Used by \code{PersonAlyticPro}.}

\item{standardize}{Logical. Should all variables be standardized? Only applies
to \code{dv}, \code{ivs}, and \code{target_ivs}.}
}
\description{
clean, a function to check data and inputs.
}
\examples{
\dontrun{
data <- clean(data=OvaryICT, ids='Mare', dv='follicles',
              time='Time', phase='Phase')
# illustrate dropping cases with < 2 observations
data <- clean(data=OvaryICT[29:nrow(OvaryICT),], ids='Mare', dv='follicles',
              time='Time', phase='Phase')
}

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
