% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{data}
\name{CurEx}
\alias{CurEx}
\title{Curelator test data}
\description{
Curelator test data
}
\examples{
# no predictor model
CurExMod <- PersonAlytic(data=CurEx,
                 ids="id",
                 dvs="severity.migraine",
                 time="time",
                 time_power=3,
                 standardize=list(dv=FALSE, iv=TRUE, byids=TRUE),
                 package="nlme")
summary(CurExMod)

# angriness predictor model
CurExModAng <- PersonAlytic(data=CurEx,
                 ids="id",
                 dvs="severity.migraine",
                 time="time",
                 ivs=list('angriness'),
                 time_power=3,
                 standardize=list(dv=FALSE, iv=TRUE, byids=TRUE),
                 package="nlme")
summary(CurExModAng)

# angriness predictor model, add correlation
CurExModAngAR1 <- PersonAlytic(data=CurEx,
                 ids="id",
                 dvs="severity.migraine",
                 time="time",
                 ivs=list('angriness'),
                 time_power=3,
                 correlation="corARMA(p=1,q=0)",
                 standardize=list(dv=FALSE, iv=TRUE, byids=TRUE),
                 package="nlme")
summary(CurExModAngAR1)

}
\author{
Stephen Tueller \email{stueller@rti.org}

Example data from Curelator
}
