% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{eds}
\alias{eds}
\title{eds - testing that an object exists}
\usage{
eds(x)
}
\arguments{
\item{x}{Any object}
}
\description{
eds - testing that an object exists
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
