% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{getParameters}
\alias{getParameters}
\title{getParameters: extract parameter table}
\usage{
getParameters(Model, package, target_iv, data, fpc)
}
\arguments{
\item{Model}{Statistical model output from \code{\link{Palytic}} methods
\code{$lme}, \code{$gamlss}, or \code{$arma}}

\item{package}{See \code{package} in \code{\link{PersonAlytic}}}

\item{data}{A data.frame, for now only pass t1$datac in \code{\link{htp}}}

\item{target_ivs}{See \code{targe_ivs} in \code{\link{PersonAlytic}}}
}
\description{
getParameters: extract parameter table
}
\keyword{internal}
