% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{htp}
\alias{htp}
\title{high throughput}
\usage{
htp(data, dims, dvs, phase, ids, uids, time, ivs, target_ivs,
  interactions = NULL, time_power = 1, correlation = NULL,
  family = gamlss.dist::NO(), standardize = list(dvs = FALSE, ivs = FALSE,
  byids = FALSE), package = "gamlss", detectAR = TRUE, PQ = c(3, 3),
  whichIC = "BIC", detectTO = TRUE, maxOrder = 3, sigma.formula = ~1,
  debugforeach = FALSE, userFormula = list(fixed = NULL, random = NULL,
  formula = NULL), cores = parallel::detectCores() - 1)
}
\description{
high throughput
}
\keyword{internal}
