% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{htpForms}
\alias{htpForms}
\title{htpForms: accumulate formula information}
\usage{
htpForms(err_id, t1, dims, package, modid)
}
\description{
htpForms: accumulate formula information
}
\keyword{internal}
