% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{monotone}
\alias{monotone}
\title{monotone}
\usage{
monotone(ids, time, data)
}
\arguments{
\item{ids}{See \code{\link{PersonAlytic}}.}

\item{time}{See \code{\link{PersonAlytic}}.}

\item{data}{See See \code{\link{PersonAlytic}}.}
}
\description{
monotone
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
