% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{data}
\name{OvaryICT}
\alias{OvaryICT}
\title{Ovary data from nlme modified for ideographic clinical trial analysis}
\description{
Ovary data from nlme modified for ideographic clinical trial analysis
}
\examples{
# A simple mixed effects models using PersonAlytic and lme
t1 <- PersonAlytic(data=OvaryICT,
                 ids="Mare",
                 dvs="follicles",
                 phase="Phase",
                 time="Time",
                 package="nlme")
summary(t1)
# Verify the PersonAlytic results to a direct call to lme
ctrl <- lmeControl(opt="optim")
t2 <- lme(follicles ~ Time * Phase + I(Time^2) * Phase,
          data = OvaryICT,
          random = ~Time + I(Time^2) | Mare,
          correlation = NULL,
          method = "REML",
          control = ctrl )
summary(t2)
# check
identical(t1,t2)
# NOTE. Time is a trigonometric function of Time, warnings are produced
# because Time is not ordered.
# definition of OvaryICT
\dontrun{
OvaryICT <- as.data.frame(nlme::Ovary)
OvaryICT$Mare <- as.numeric(OvaryICT$Mare)
OvaryICT$Phase <- as.numeric(OvaryICT$Time > .5)
OvaryICT$Time <- sin(2*pi*OvaryICT$Time)
OvaryICT <- OvaryICT[order(OvaryICT$Mare),]
}

# output verification of t1
capture.output(t1.coef <- summary(t1), file = 'NUL')
all.equal(as.vector(t1.coef[,1]),
c(11.2421402, -1.9730995, 0.9126069, -1.9746526, 1.0967642))
}
\references{
Pinero, J., & Bates, D. (2000). Mixed-effects models in S and S-PLUS
(statistics and computing).
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
