% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{data}
\name{OvaryICT}
\alias{OvaryICT}
\title{Ovary data from nlme modified for ideographic clinical trial analysis}
\description{
Ovary data from nlme modified for ideographic clinical trial analysis
}
\examples{
# A simple mixed effects models using PersonAlytic and lme
t1 <- PersonAlytic(data=OvaryICT,
                   ids="Mare",
                   dvs="follicles",
                   phase="Phase",
                   time="Time",
                   package="nlme",
                   detectAR = FALSE,
                   detectTO = FALSE,
                   standardize = FALSE,
                   alignPhase = FALSE)

# Verify the PersonAlytic results to a direct call to lme
t2 <- lme(follicles ~ Time * Phase,
          data = OvaryICT,
          random = ~Time | Mare,
          method = "REML",
          control = t1$call$control)
summary(t2)

# verification tests
all.equal(summary(t1)$tTable, summary(t2)$tTable)
all.equal(VarCorr(t1), VarCorr(t2))

# verify estimates
all.equal( summary(t1)$tTable[,1],
c(`(Intercept)`  = 10.6616306,
   Time          = -0.8689801,
   Phase         = 10.9720945,
   `Time:Phase`  = -8.6438502) )

# definition of OvaryICT
\dontrun{
OvaryICT <- as.data.frame(nlme::Ovary)
OvaryICT$Mare <- as.numeric(OvaryICT$Mare)
OvaryICT$Phase <- as.numeric(OvaryICT$Time > .5)
OvaryICT$Time <- sin(2*pi*OvaryICT$Time)
OvaryICT <- OvaryICT[order(OvaryICT$Mare),]
}

}
\references{
Pinero, J., & Bates, D. (2000). Mixed-effects models in S and S-PLUS
(statistics and computing).
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
