% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlytics.r
\docType{data}
\name{OvaryICT}
\alias{OvaryICT}
\title{Ovary data from nlme modified for ideographic clinical trial analysis}
\description{
Ovary data from nlme modified for ideographic clinical trial analysis
}
\examples{
# A simple mixed effects models using PersonAlytic and lme
t1 <- PersonAlytic(data=OvaryICT,
                   ids="Mare",
                   dvs="follicles",
                   phase="Phase",
                   time="Time",
                   package="nlme",
                   detectAR = FALSE,
                   detectTO = FALSE,
                   standardize = list(dvs=FALSE,ivs=FALSE,byids=FALSE),
                   alignPhase = FALSE)
summary(t1)

# Verify the PersonAlytic results to a direct call to lme
t2 <- lme(follicles ~ Time * Phase,
          data = OvaryICT,
          random = ~Time | Mare,
          method = "REML",
          control = t1$call$control)
summary(t2)

# verification tests - are the results the same?
all.equal(summary(t1)$tTable, summary(t2)$tTable)
all.equal(VarCorr(t1), VarCorr(t2))

# verify estimates against known values
all.equal( summary(t1)$tTable[,1],
c(`(Intercept)`  = 11.2100153,
   Time          = -1.9504839,
   Phase         = 0.9115449,
   `Time:Phase`  = -1.9746812) )

# Illustrate a multiphase study, Phase alignment (To do: add visualization)
t3 <- PersonAlytic(data=OvaryICT,
                   ids="Mare",
                   dvs="follicles",
                   phase="Phase2",
                   time="Time",
                   package="nlme",
                   detectAR = FALSE,
                   detectTO = FALSE,
                   standardize = list(dvs=FALSE,ivs=FALSE,byids=FALSE),
                   alignPhase = TRUE)
summary(t3)

# definition of OvaryICT
\dontrun{
OvaryICT <- as.data.frame(nlme::Ovary)
OvaryICT$Mare <- as.numeric(OvaryICT$Mare)
OvaryICT$Phase <- as.numeric(OvaryICT$Time > .5)
OvaryICT$Phase2 <- factor( cut(OvaryICT$Time,
                               breaks = c(-Inf, .2, .6, 1, Inf)),
                           labels = 1:4 )
OvaryICT$Time <- sin(2*pi*OvaryICT$Time)
set.seed(1)
t1 <- matrix( sample(1:4, 3*nrow(OvaryICT), replace = TRUE), ncol = 3)
t1 <- data.frame(t1)
names(t1) <- paste('Target', 1:3, sep='')
set.seed(2)
t2 <- data.frame( matrix( rnorm(3*nrow(OvaryICT)), ncol = 3 ) )
names(t2) <- paste('Target', 4:6, sep='')
OvaryICT <- data.frame(OvaryICT, t1, t2)
OvaryICT <- OvaryICT[order(OvaryICT$Mare),]
}

}
\references{
Pinero, J., & Bates, D. (2000). Mixed-effects models in S and S-PLUS
(statistics and computing).
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
