% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palytic.r
\docType{class}
\name{Palytic}
\alias{Palytic}
\title{\code{Palytic} class generator.}
\format{A \code{Palytic} generator}
\usage{
Palytic(data, ids, dv, time)
}
\description{
\code{Palytic} class generator.
}
\details{
The fields \code{data}, \code{ids}, \code{dv}, and \code{time} are required.
Using these, the default model \eqn{dv=time} with random intercepts for \code{ids}
and random intercepts for \code{time} is constructed. See the example. If
\code{phase} is provided, the default model is \eqn{dv=time+phase+phase*time}, and if
\code{ivs} are provided they are included in the model.
}
\section{Fields}{

\describe{
\item{\code{data}}{A \code{\link{data.frame}} that contains as variables \code{ids},
\code{dv}, \code{phase}, and \code{time}. Optionally, additional independent
variables can be included in \code{ivs}. \code{fixed} and \code{random} formulae
for \code{\link{lme}} models and \code{formula} for \code{\link{gamlss}} models are
automatically generated when a \code{Palytic} object is created if these fields
are left \code{NULL}.}

\item{\code{ids}}{A character string giving the name of the id variable in \code{data}.}

\item{\code{dv}}{A character string giving the name of the dependent variable in \code{data}.}

\item{\code{time}}{A character string giving the name of the time variable in \code{data}.
Random slopes for time are inclued by default. This can be overridden by specifying
\code{fixed} and \code{random} formula for \code{\link{lme}} models or by specifying
the \code{formula} for \code{\link{gamlss}} models.}

\item{\code{phase}}{A character string giving the name of the phase variable in \code{data}.
The \code{phase*time} interaction is included by default. This can be overridden by
specifying \code{fixed} and \code{random} formula for \code{\link{lme}} models or by
specifying the \code{formula} for \code{\link{gamlss}} models.}

\item{\code{ivs}}{A \code{\link{list}} of one or more character strings giving the names
of additional variables in \code{data}, e.g., \code{list('iv2', 'iv2')}.}

\item{\code{interactions}}{List of vector pairs of variable names for which interaction
terms should be specified, e.g., \code{list(c('time', 'phase'), c('time', 'iv1'),
c('iv1', 'iv2'))} where \code{'iv1'} is the name of a variable in the liste \code{ivs}.}

\item{\code{time_power}}{The polynomial for \code{time}, e.g., \code{time^time_power}. Fixed
effects for \code{time^1...time^time_power} will be included in models. Future
releases will allow for other functions of time such as \code{\link{sin}}, but these
can be applied directly by transforming the \code{time} variable.}

\item{\code{correlation}}{See \code{\link{corStruct}}. Defaults to \code{NULL}, see
\code{\link{lme}}. Used by both \code{\link{lme}} and \code{\link{gamlss}} models.}

\item{\code{family}}{The \code{\link{gamlss.family}} distribution.}

\item{\code{fixed}}{The \code{fixed} effects model for \code{\link{lme}} models.}

\item{\code{random}}{The \code{random} effects model for \code{\link{lme}} models.}

\item{\code{formula}}{The \code{formula} effects model for \code{\link{gamlss}} models.
\code{sigma.formula}, \code{nu.formula}, and \code{tau.formula} will be implemented in
a future release.}

\item{\code{method}}{See \code{method} in \code{\link{lme}}. Is usef for both \code{\link{lme}}
and \code{\link{gamlss}} models.}

\item{\code{standardize}}{Logical. Should \code{y} and \code{ivs} be standardized?}

\item{\code{corStructs}}{Vector. A \code{correlation} structure for each case in \code{ids}. Not
user accesible. Populated by \code{\link{PersonAylticsPro}}.}

\item{\code{time_powers}}{Vector. A \code{time_order} for each case in \code{ids}. Not
user accesible. Populated by \code{\link{PersonAylticsPro}}.}

\item{\code{is_clean}}{Logical flag for data cleaning. Not yet implemented.}

\item{\code{warnings}}{A list of warnings that will be populated as methods are called on a
\code{Palytic} object.}

\item{\code{errors}}{A list of errors that will be populated as methods are called on a
\code{Palytic} object.}

\item{\code{try_silent}}{Logical flag for testing error handling in \code{Palytic} methods.}
}}

\section{Methods}{

\describe{
  \item{\code{lme(subgroup = NULL)}}{This method fits the linear mixed effects
   \code{lme} model implied by the \code{Palytic} fields \code{ids},
   \code{dv}, \code{phase}, \code{time}, and optionally \code{ivs}, \code{time_power}
   \code{correlation}. The default formula can be overridden by specifying
   \code{fixed} and \code{random} and optionally \code{method} and \code{correlation}.
   The see \code{\link{lme}}
  for the parameter \code{subgroup}. }
  \item{\code{gamlss(subgroup = NULL)}}{This method fits the \code{lme} model implied
  by the \code{Palytic} fields \code{ids}, \code{dv}, \code{phase}, \code{time}
  and optionally \code{ivs}, \code{time_power}, \code{correlation}. The default
  formula can be overridden by specifying \code{formula} and
  optionally \code{method}, \code{correlation}, and \code{family} (which can be
  used to specify generalized linear mixed effects models,
  see \code{\link{gamlss.family}}).
  The parameter \code{subgroup} operates as in \code{\link{lme}}.}
  #'   \item{\code{getAR_order(dv, maxAR=3, maxMA=3, crit="BIC", lrt=TRUE, alpha=.05)}}{
  This method automates the task of determining the correlation structure for each case in
  \code{ids} (see \code{\link{PersonAlytic} or \code{\link{PersonAlytic}}}).
  \code{maxAR} and \code{maxMA} set the highest autoregressive and moving
  average parameters to be tested. If the time variable is approximatetly equally spaced,
  \code{crit} is the criterion used for determining the correlation structure for each
  \code{ids} using the \code{\link{auto.arima}} function. If the time variable is unequally
  spaced, \code{crit} as also the criterion for
  model selection via mixed effects models using \code{\link{lme}} if \code{lrt=FALSE}.
  If \code{lrt=TRUE} (the default), likelihood ratios are used via the \code{\link{anova}}
  method for \code{\link{lme}} objects. Calling \code{getAR_order} populates the
  \code{corStructs} field of a \code{Palytic} object. For usage, see the examples.}
  \item{\code{getTime_Power(subset, maxOrder)}}{This method automates the task of
  determining  \code{time_power} for each case in \code{ids}
  (see \code{\link{PersonAlytic} or \code{\link{PersonAlytic}}}). For example,
  if \code{getTime_Power} returns \code{time_power=3}, then \code{time + time^2 + time^3}
  will be added to the fixed effects of the model.
  Calling \code{getTime_Power} populates the
  \code{time_powers} field of a \code{Palytic} object. For usage, see the examples.}
}
}

\examples{
# construct a new Payltic object and examine the default formulae
t1 <- Palytic$new(data = OvaryICT, ids='Mare', dv='follicles',
                  time='Time', phase='Phase')
t1$fixed
t1$random
t1$formula
t1$correlation
# compare gamlss and lme output, in which the models of the default formulae are fit
(t1.gamlss <- summary(t1$gamlss()))
(t1.lme    <- summary(t1$lme()))
# parameter estimates are equal within 0.01
all.equal(t1.gamlss[1:4,1], t1.lme$tTable[,1], tolerance = 0.01)
# now change the correlation structure and compare gamlss and lme output,
# noting that the intercepts are very different now
t1$correlation <- "corARMA(p=1, q=0)"
summary(t1$gamlss())
summary(t1$lme())
# fit the model only to the first mare with ML instead of REML
t1$method <- 'ML'
summary(t1$gamlss(OvaryICT$Mare==1))
# quadratic model
t2 <- t1$clone()
#t2$random <- formula("Time + I(Time^2) | Mare") # debug, this doesn't work
#t2$lme()
# change the time variable and notice perpetuate through the other objects
t1$time <- 'Time'
t1$fixed
t1$random
t1$formula
# change the formula (note limitations with the interaciton)
t2 <- t1$clone()
t2$formula <- formula(follicles ~ Time * Phase +
                      re(random = ~Time + I(Time^2) | Mare, method = "ML",
                      correlation = corARMA(p=1,q=0)))
t2$formula

t1 <- Palytic$new(data = PersonAlyticsLite::OvaryICT, ids='Mare',
                  dv='follicles', time='TimeSin', phase='Phase')
t1$getTime_Power()
t1$time_powesr
# getAR_order works on one case at a time
t1$getAR_order(t1$dv)
t1$corStructs
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{data}
