% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palytic.r
\docType{class}
\name{Palytic}
\alias{Palytic}
\title{\code{Palytic} class generator.}
\format{A \code{Palytic} generator}
\usage{
Palytic(data, ids, dv, time)
}
\description{
\code{Palytic} class generator.
}
\details{
The fields \code{data}, \code{ids}, \code{dv}, and \code{time} are required.
Using these, the default model \eqn{dv=time} with random intercepts for \code{ids}
and random intercepts for \code{time} is constructed. See the example. If
\code{phase} is provided, the default model is \eqn{dv=time+phase+phase*time}, and if
\code{ivs} are provided they are included in the model.
}
\section{Fields}{

\describe{
\item{\code{data}}{A \code{\link{data.frame}} that contains as variables \code{ids},
\code{dv}, \code{phase}, and \code{time}. Optionally, additional independent
variables can be included in \code{ivs}. \code{fixed} and \code{random} formulae
for \code{\link{lme}} models and \code{formula} for \code{\link{gamlss}} models are
automatically generated when a \code{Palytic} object is created if these fields
are left \code{NULL}.}

\item{\code{ids}}{A character string giving the name of the id variable in \code{data}.}

\item{\code{dv}}{A character string giving the name of the dependent variable in \code{data}.}

\item{\code{time}}{A character string giving the name of the time variable in \code{data}.
Random slopes for time are inclued by default. This can be overridden by specifying
\code{fixed} and \code{random} formula for \code{\link{lme}} models or by specifying
the \code{formula} for \code{\link{gamlss}} models.}

\item{\code{phase}}{A character string giving the name of the phase variable in \code{data}.
The \code{phase*time} interaction is included by default. This can be overridden by
specifying \code{fixed} and \code{random} formula for \code{\link{lme}} models or by
specifying the \code{formula} for \code{\link{gamlss}} models.}

\item{\code{ivs}}{A \code{\link{list}} of one or more character strings giving the names
of additional variables in \code{data}, e.g., \code{list('iv2', 'iv2')}.}

\item{\code{interactions}}{List of vector pairs of variable names for which interaction
terms should be specified, e.g., \code{list(c('time', 'phase'), c('time', 'iv1'),
c('iv1', 'iv2'))} where \code{'iv1'} is the name of a variable in the liste \code{ivs}.}

\item{\code{time_power}}{The polynomial for \code{time}, e.g., \code{time^time_power}. Fixed
effects for \code{time^1...time^time_power} will be included in models. Future
releases will allow for other functions of time such as \code{\link{sin}}, but these
can be applied directly by transforming the \code{time} variable.}

\item{\code{correlation}}{See \code{\link{corStruct}}. Defaults to \code{NULL}, see
\code{\link{lme}}. Used by both \code{\link{lme}} and \code{\link{gamlss}} models.}

\item{\code{family}}{The \code{\link{gamlss.family}} distribution.}

\item{\code{fixed}}{The \code{fixed} effects model for \code{\link{lme}} models.}

\item{\code{random}}{The \code{random} effects model for \code{\link{lme}} models.}

\item{\code{formula}}{The \code{formula} effects model for \code{\link{gamlss}} models.
\code{sigma.formula}, \code{nu.formula}, and \code{tau.formula} will be implemented in
a future release.}

\item{\code{method}}{See \code{method} in \code{\link{lme}}. Is usef for both \code{\link{lme}}
and \code{\link{gamlss}} models.}

\item{\code{standardize}}{Named logical list. Which variables should be standardized? The default
is \code{list(dv=FALSE, ivs=FALSE, byids=FALSE)}. See \code{dv} and \code{ivs}. The option
\code{byids} controls whether standardization is done by individuals or by group. Any time
variables are changed (e.g., \code{ivs}), the data are subset, or the options in
\code{standardize} are changed, the raw data will be restandardized (see \code{datac}).}

\item{\code{corStructs}}{Vector. A \code{correlation} structure for each case in \code{ids}. Not
user accesible. Populated by \code{\link{PersonAlytic}}.}

\item{\code{time_powers}}{Vector. A \code{time_order} for each case in \code{ids}. Not
user accesible. Populated by \code{\link{PersonAlytic}}.}

\item{\code{datac}}{data.frame. Cleaned data. Cleaning involves the following steps:
1. Check that the variables in \code{ids}, \code{dv}, \code{time}, \code{phase},
\code{ivs}, and \code{interactions} are in \code{data}.
2. The \code{ids} variable is forced to be numeric.
3. The validity of the \code{correlation} structure is checked, see \code{\link{corStruct}}.
4. check that variables have non-zero variance.
5. If standardization is requested, standardize the data (see \code{standardize}).
6. Sort the data on \code{ids} and \code{time}.
7. If patients have < 2 observations, they are dropped from the data set.
8. If phase alignment is requested, align phases (see \code{alignPhases}).}

\item{\code{warnings}}{A list of warnings that will be populated as methods are called on a
\code{Palytic} object.}

\item{\code{errors}}{A list of errors that will be populated as methods are called on a
\code{Palytic} object.}

\item{\code{try_silent}}{Logical flag for testing error handling in \code{Palytic} methods.}
}}

\section{Methods}{

\describe{

  \item{\code{summary}}{This method provides a summary of the inputs, the cleaned data,
  and the raw data.}

  \item{\code{describe}}{This method gives the correlation between \code{dv} and
  each continuous variable in \code{ivs} (as well as the \code{time} variable), or,
  if variablse are factors (including \code{phase}), the mean of \code{dv} is given
  for each factor level of each variable.}

  \item{\code{lme(subgroup = NULL, dropVars = NULL)}}{This method fits the
  linear mixed effects
   \code{lme} model implied by the \code{Palytic} fields \code{ids},
   \code{dv}, \code{phase}, \code{time}, and optionally \code{ivs},
   \code{time_power}
   \code{correlation}. The default formula can be overridden by specifying
   \code{fixed} and \code{random} and optionally \code{method} and
   \code{correlation}. The see \code{\link{lme}} for the parameter
   \code{subgroup}. The \code{dropVars} parameter indicates which fixed
   effects should be dropped for a likelihood ration test (LRT). This is
   used by \code{\link{PersonAlytic}} to test \code{target_ivs}.}

  \item{\code{gamlss(subgroup = NULL, sigma.formula = ~1, family = NULL,
  dropVars = NULL)}}{This method fits the \code{lme} model implied
  by the \code{Palytic} fields \code{ids}, \code{dv}, \code{phase}, \code{time}
  and optionally \code{ivs}, \code{time_power}, \code{correlation}. The default
  formula can be overridden by specifying \code{formula} and
  optionally \code{method}, \code{correlation}, and \code{family} (which can be
  used to specify generalized linear mixed effects models,
  see \code{\link{gamlss.family}}).
  The parameter \code{subgroup} operates as in \code{\link{lme}}. The parameter
  \code{sigma.formula} and \code{family} are desribed in \code{\link{gammlss}}.
   The \code{dropVars} parameter indicates which fixed
   effects should be dropped for a likelihood ration test (LRT). This is
   used by \code{\link{PersonAlytic}} to test \code{target_ivs}.}

  \item{\code{arma(subgroup=NULL, max.p=3, max.q=3, dropVars=NULL,
  max.P=0, max.Q=0, max.d=0, max.D=0)}}{For individual level models,
  random intercepts and
  random slopes are not defined. In this situatation, an \code{ARMA(p,q)}
  should be used. This is implemented using the \code{xreg}
  option of \code{\link{arima}}. The residual correlation search is achieved using
  \code{\link{auto.arima}}.  The \code{dropVars} parameter indicates which fixed
   effects (in \code{xreg}) should be dropped for a likelihood ration test
   (LRT). This is used by \code{\link{PersonAlytic}} to test \code{target_ivs}.
  The other parameters are as in \code{\link{auto.arima}}.}

  \item{\code{getAR_order(P=3, Q=3, whichIC="BIC", lrt=FALSE, alpha=.05)}}{
  --- slated for deprication, superceded by \code{arma}. ---
  This method automates the task of determining the correlation structure for
  each case in \code{ids} (see \code{\link{PersonAlytic} or
  \code{\link{PersonAlytic}}}). \code{P} and \code{Q} set the highest
  autoregressive and moving average parameters to be tested. If the time
  variable is approximatetly equally spaced, \code{whichIC} is the criterion
  used for determining the correlation structure for each \code{ids} using
  the \code{\link{auto.arima}} function. If the time variable is unequally
  spaced, \code{whichIC} as also the criterion for model selection via mixed
  effects models using \code{\link{lme}} if \code{lrt=FALSE}. If
  \code{lrt=TRUE} likelihood ratios are used via the \code{\link{anova}}
  method for \code{\link{lme}} objects. This is NOT reccomended unless Q=0
  and only AR models are considered since AR and MA models are not nested.
  Calling \code{getAR_order} populates the \code{corStructs} field of a
  \code{Palytic} object. For usage, see the examples.}

  \item{\code{GroupAR_order(dv, P=3, Q=3, whichIC="BIC", lrt=FALSE, alpha=.05)}}{The
  same as \code{getAR_order} when the ARMA order is desired for the full sample.}
  \item{\code{getTime_Power(subset, maxOrder)}}{This method automates the task of
  determining  \code{time_power} for each case in \code{ids}
  (see \code{\link{PersonAlytic} or \code{\link{PersonAlytic}}}). For example,
  if \code{getTime_Power} returns \code{time_power=3},
  then \code{time + time^2 + time^3}
  will be added to the fixed effects of the model.
  Calling \code{getTime_Power} populates the
  \code{GroupTime_power} field of a \code{Palytic} object. For usage, see the examples.}

  \item{\code{groupTime_Power(subset, maxOrder)}}{The same as \code{getTime_power} when
  the polynomial of time is desired for the full sample.}
}
}

\examples{
# construct a new Payltic object and examine the default formulae
t1 <- Palytic$new(data = OvaryICT, ids='Mare', dv='follicles',
                  time='Time', phase='Phase')
t1$fixed
t1$random
t1$formula
t1$correlation

# summary function
t1$summary()

# descriptive statistics
t1$describe()

# compare gamlss and lme output, in which the models of the default formulae are fit
(t1.gamlss <- summary(t1$gamlss()))
(t1.lme    <- summary(t1$lme()))

# parameter estimates are equal within 0.01
message('\\n\\nAre `lme` and `gamlss` results equivalent?\\n',
all.equal(t1.gamlss[1:4,1], t1.lme$tTable[,1], tolerance = 0.01) )

# now change the correlation structure and compare gamlss and lme output,
# noting that the intercepts are very different now
t1$correlation <- "corARMA(p=1, q=0)"
summary(t1$gamlss())
summary(t1$lme())

# fit the model only to the first mare with ML instead of REML
t1$method <- 'ML'
summary(t1$gamlss(OvaryICT$Mare==1))

# change the formula
t2 <- t1$clone()
t2$formula <- formula(follicles ~ Time * Phase +
                      re(random = ~Time + I(Time^2) | Mare, method = "ML",
                      correlation = corARMA(p=1,q=0)))
t2$formula

# random intercept only model
t2 <- t1$clone()
t2$random <- formula(~1|Mare)
t2$random
t2$formula

# random slope only model
t2$random <- formula(~0+Time|Mare)
t2$random
t2$formula

# automatically select the polynomial order of time with getTime_Power
t1 <- Palytic$new(data = OvaryICT, ids='Mare',
                  dv='follicles', time='Time', phase='Phase')
t1$getTime_Power()
t1$time_powers

# automatically select the ARMA model for residual correlation getAR_order
# this runs slow because a model is selected for each individual
\dontrun{
t1$getAR_order()
t1$corStructs
}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{data}
