% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pwtime}
\alias{pwtime}
\title{pwtime - function to make piecewise time variables for a multiphase study}
\usage{
pwtime(time, phase)
}
\arguments{
\item{time}{Numeric vector. A sequence of study times. If any times are
negative (e.g., if time is center at 0 between the first and second
phase) it is rescale to have a minimum of 0 so that the first obesravation
within each phase has time=0.}

\item{phase}{Numeric or character vector of the same length as time. The
phases within which time is rescale to creat the time variables needed for
a piecewise growth model.}
}
\description{
pwtime - function to make piecewise time variables for a multiphase study
}
\examples{

pwtime(-49:50, sort(rep(letters[1:5], length.out = 100)))

# time should not be scaled in [0,1] or similar, rescale to integer
OvaryICT2 <- OvaryICT
OvaryICT2$Time <- round(OvaryICT2$Time*30)

p1 <- Palytic$new(data = OvaryICT2, ids='Mare', dv='follicles',
                  time='Time', phase='Phase', alignPhase = 'none')

p2 <- Palytic$new(data = OvaryICT2, ids='Mare', dv='follicles',
                  time='Time', phase='Phase', alignPhase = 'piecewise')

t1 <- PersonAlytic(data=OvaryICT2                                      ,
                   ids="Mare"                                          ,
                   dvs="follicles"                                     ,
                   phase="Phase"                                       ,
                   time="Time"                                         ,
                   package="nlme"                                      ,
                   detectAR = FALSE                                    ,
                   detectTO = FALSE                                    ,
                   standardize = list(dvs=FALSE,ivs=FALSE,byids=FALSE) ,
                   alignPhase = 'none'                                 )

t2 <- PersonAlytic(data=OvaryICT2                                      ,
                   ids="Mare"                                          ,
                   dvs="follicles"                                     ,
                   phase="Phase"                                       ,
                   time="Time"                                         ,
                   package="nlme"                                      ,
                   detectAR = FALSE                                    ,
                   detectTO = FALSE                                    ,
                   standardize = list(dvs=FALSE,ivs=FALSE,byids=FALSE) ,
                   alignPhase = 'piecewise'                            )
}
\author{
Stephen Tueller \email{Stueller@rti.org}
}
