% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htp.r
\name{rcm}
\alias{rcm}
\title{rcm: function to force values to data.frame vectors for later stacking}
\usage{
rcm(x, target_ivs)
}
\description{
rcm: function to force values to data.frame vectors for later stacking
}
\keyword{internal}
